Ensure destroy is idempotent
Unfortunately the graph driver doesn't make it easy to distinguish
between errors because a layer is already gone and errors because we
couldn't figure out if a layer is gone. In the case where the layer
can't be retrieved but is still there this will leak a layer (and log
loudly about it), but this should be extremely uncommon, and will be
caught by Garbage Collection when enabled.

Signed-off-by: Julian Friedman <julz.friedman@uk.ibm.com>